Techdebt: rename stores file in nasdaqapi
